feat: Phase 8 — AI Copilot module with agentic LLM chat
Add M18 AI Copilot tab with 12 agentic tools for querying market data,
portfolio, options, fundamentals, and economic indicators. Supports
Claude API (primary) and Ollama (local fallback).

New files:
- data/llm_agent.py: Agent logic, tool definitions, provider abstraction
- modules/m18_ai_copilot.py: Bloomberg-themed chat UI with suggestions

Also fixes:
- m16_alerts: StreamlitMixedNumericTypesError on number_input (int/float)

+++ modules/m16_alerts.py
@@ -119,7 +119,7 @@ def _render_add_alert():
 
     with c_val:
         # Default value based on LTP if available
-        default_price = round(_quote["lastPrice"], 2) if _quote else 100.0
+        default_price = float(round(_quote["lastPrice"], 2)) if _quote else 100.0
         if condition in ("pct_change_above", "pct_change_below"):
             value = st.number_input(
                 "VALUE (%)",
